requirements.txt
requests
beautifulsoup4
langchain==0.2.14
langchain-huggingface==0.0.3
langchain-google-genai==1.0.8
chromadb==0.5.5
langchain-chroma==0.1.2
sentence-transformers==3.0.1
python-dotenv==1.0.1
